refactor: TestDataCreator -> creator/TestUserCreator
